Create PipelineApiController
