+++ src/main/java/gov/cdc/izgateway/xform/ApiController.java
@@ -26,7 +26,6 @@ import java.util.logging.Level;
 @Lazy(false)
 public class ApiController extends BaseApiController {
     private final OrganizationService organizationService;
-    private final PipelineService pipelineService;
     private final SolutionService solutionService;
     private final MappingService mappingService;
     private final PreconditionService preconditionService;
@@ -41,7 +40,6 @@ public class ApiController extends BaseApiController {
     @Autowired
     public ApiController(
             OrganizationService organizationService,
-            PipelineService pipelineService,
             SolutionService solutionService,
             MappingService mappingService,
             PreconditionService preconditionService,
@@ -52,7 +50,6 @@ public class ApiController extends BaseApiController {
             AccessControlRegistry registry
     ) {
         this.organizationService = organizationService;
-        this.pipelineService = pipelineService;
         this.solutionService = solutionService;
         this.mappingService = mappingService;
         this.preconditionService = preconditionService;
@@ -61,24 +58,6 @@ public class ApiController extends BaseApiController {
         this.groupRoleMappingService = groupRoleMappingService;
         this.accessControlService = accessControlService;
         registry.register(this);
-    }
-
-    @RolesAllowed({Roles.PIPELINE_READER})
-    @GetMapping("/api/v1/pipelines/{uuid}")
-    public ResponseEntity<Pipeline> getPipelineByUUID(@PathVariable UUID uuid) {
-        Pipeline entity = pipelineService.getObject(uuid);
-        if (entity == null) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
-        return new ResponseEntity<>(entity, HttpStatus.OK);
-    }
-
-    @RolesAllowed({Roles.PIPELINE_WRITER})
-    @PutMapping("/api/v1/pipelines/{uuid}")
-    public ResponseEntity<Pipeline> updatePipeline(@PathVariable UUID uuid, @Valid @RequestBody Pipeline updatedPipeline) {
-        updatedPipeline.setId(uuid);
-        pipelineService.update(updatedPipeline);
-        return new ResponseEntity<>(updatedPipeline, HttpStatus.OK);
     }
 
     @RolesAllowed({Roles.SOLUTION_READER})
@@ -115,23 +94,6 @@ public class ApiController extends BaseApiController {
         updatedOrganization.setId(uuid);
         organizationService.update(updatedOrganization);
         return new ResponseEntity<>(updatedOrganization, HttpStatus.OK);
-    }
-
-    // TODO - Paul to look at AccessControl as it works with groups too.
-    @RolesAllowed({Roles.PIPELINE_READER})
-    @GetMapping("/api/v1/pipelines")
-    public ResponseEntity<String> getPipelinesList(
-            @RequestParam(required = false) String nextCursor,
-            @RequestParam(required = false) String prevCursor,
-            @RequestParam(defaultValue = "false") Boolean includeInactive,
-            @RequestParam(defaultValue = "10") int limit
-    ) {
-        try {
-            return processList(pipelineService.getList(), nextCursor, prevCursor, includeInactive, limit);
-        } catch (JsonProcessingException e) {
-            log.log(Level.SEVERE, e.getMessage(), e);
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
-        }
     }
 
     @RolesAllowed({Roles.ORGANIZATION_READER})
@@ -343,15 +305,6 @@ public class ApiController extends BaseApiController {
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
-    @RolesAllowed({Roles.PIPELINE_WRITER})
-    @PostMapping("/api/v1/pipelines")
-    public ResponseEntity<Pipeline> createPipeline(
-            @Valid @RequestBody() Pipeline pipeline
-    ) {
-        pipelineService.create(pipeline);
-        return new ResponseEntity<>(pipeline, HttpStatus.OK);
-    }
-
     @RolesAllowed({Roles.ORGANIZATION_WRITER})
     @PostMapping("/api/v1/organizations")
     public ResponseEntity<Organization> createOrganization(
@@ -368,20 +321,6 @@ public class ApiController extends BaseApiController {
     ) {
         solutionService.create(solution);
         return new ResponseEntity<>(solution, HttpStatus.OK);
-    }
-
-    @RolesAllowed({Roles.PIPELINE_DELETER})
-    @DeleteMapping("/api/v1/pipelines/{uuid}")
-    public ResponseEntity<Pipeline> deletePipeline(
-            @PathVariable UUID uuid
-    ) {
-
-        if (Boolean.FALSE.equals(allowDelete)) {
-            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
-        }
-
-        pipelineService.delete(uuid);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
     @RolesAllowed({Roles.SOLUTION_DELETER})
